removed old error handling buttons and added dependency on NnSwiftUIErrorHandling package
