+++ Package.swift
@@ -10,13 +10,15 @@ let package = Package(
         .library(name: "NnLoginKit", targets: ["NnLoginKit"])
     ],
     dependencies: [
-        .package(url: "https://github.com/google/GoogleSignIn-iOS.git", from: "7.0.0")
+        .package(url: "https://github.com/google/GoogleSignIn-iOS.git", from: "7.0.0"),
+        .package(url: "https://github.com/nikolainobadi/NnSwiftUIErrorHandling.git", from: "1.0.0"),
     ],
     targets: [
         .target(
             name: "NnLoginKit",
             dependencies: [
-                .product(name: "GoogleSignIn", package: "GoogleSignIn-iOS")
+                .product(name: "GoogleSignIn", package: "GoogleSignIn-iOS"),
+                .product(name: "NnSwiftUIErrorHandling", package: "NnSwiftUIErrorHandling"),
             ],
             path: "Sources",
             resources: [
@@ -29,4 +31,3 @@ let package = Package(
         ),
     ]
 )
-
