+++ iOSLoginSample/iOSLoginSample.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -35,6 +35,15 @@
         "revision" : "cee3c709307912d040bd1e06ca919875a92339c6",
         "version" : "2.0.0"
       }
+    },
+    {
+      "identity" : "nnswiftuierrorhandling",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/nikolainobadi/NnSwiftUIErrorHandling.git",
+      "state" : {
+        "revision" : "b04c633162720e1cda30ea1ae8cd60cd62cd5594",
+        "version" : "1.0.0"
+      }
     }
   ],
   "version" : 2
